Correct Precedence and Create Operate Function
Correct the function that determines operator precedence to add the current operator to the stack if the stack is empty.

Create the operate function which carries out the given operator's associated operation on the two top values in the value stack. Replaces the penultimate value with the result and removes the last value. Also removes the operator from the stack after execution.